+++ internal/bootstrap/config/loadConfiguration.go
@@ -1,4 +1,4 @@
-package bootstrap
+package config
 
 import (
 	"encoding/json"
